refactor: move RLWorkspaceManager to correct dir
