+++ src/agentscope/app/workspace_manager/__init__.py
@@ -3,13 +3,15 @@
 and their lifecycles, and filesystem isolation."""
 
 from ._base import WorkspaceManagerBase
-from ._local_workspace_manager import LocalWorkspaceManager
 from ._docker_workspace_manager import DockerWorkspaceManager
 from ._e2b_workspace_manager import E2BWorkspaceManager
+from ._local_workspace_manager import LocalWorkspaceManager
+from ._rl_workspace_manager import RLWorkspaceManager
 
 __all__ = [
     "WorkspaceManagerBase",
     "LocalWorkspaceManager",
     "DockerWorkspaceManager",
     "E2BWorkspaceManager",
+    "RLWorkspaceManager",
 ]
