+++ src/agentscope/app/workspace_manager/_rl_workspace_manager.py
@@ -39,7 +39,7 @@ from ...workspace._e2b._bootstrap import (
     DEFAULT_TEMPLATE,
     DEFAULT_TIMEOUT,
 )
-from ..workspace_manager._base import WorkspaceManagerBase
+from ._base import WorkspaceManagerBase
 from ._workspace_pool import PooledEntry, WorkspacePool
