+++ src/agentscope/app/workspace_manager/_workspace_pool.py
